structure changes to paymentflow
