+++ autopayroll-CordaService/workflows-java/src/test/java/net/corda/examples/autopayroll/FlowTests.java
@@ -67,5 +67,5 @@ public class FlowTests {
                     assertEquals(ptx.getId(), txHash);
                 });
     }
-    
+
 }
